requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.4.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.3
django-allauth==0.43.0
django-countries==6.1.3
django-crispy-forms==1.9.2
django-filter==2.4.0
django-heroku==0.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.2
filelock==3.0.12
Flask==1.1.2
flask-mongoengine==0.9.5
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.20.0
oauthlib==3.1.0
Pillow==8.0.1
pipenv==2020.8.13
psycopg2==2.8.6
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
pylint==2.6.0
pymongo==3.11.0
python-decouple==3.3
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2019.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.14.0
source==1.2.0
sqlparse==0.3.0
stripe==2.55.0
toml==0.10.1
urllib3==1.25.11
virtualenv==20.0.4
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.5
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
WTForms==2.3.3
